condor_annex version 0.0.1
Commented out call to the subroutine waitForNodesInPool in the main
condor_annex script. This subroutine does not account for the existence
of other HTCondor pools and the potential flocking of jobs between them,
where the annex instances may be connected to another central manager.
It assumes COLLECTOR_HOST on the localhost executing condor_annex is
the annex central manager. In the future, waitForNodesInPool should be
modified to use the condor_status -pool option to explicitly query the
correct COLLECTOR_HOST. However, any such modification should also
attempt to take into account whether or not the COLLECTOR_HOST may be
different depending on which daemon is trying to contact the central
manager. For example, this may occur when the annex instances contact
the central manager via a private IP address, while the submit node
contacts it via a public IP address.